GQL: Extracting files from GQLSchema to models.gql
Still need to refactor GQLTissue.
